+++ _modules/_slider/slider.js
@@ -1,5 +1,6 @@
 var $ = require('jquery');
 var Slick = require('../../node_modules/slick-carousel/slick/slick');
+var Tracking = require('../_tracking/tracking');
 
 $('[data-slider]').each(function () {
     $(this).slick({
@@ -17,5 +18,13 @@ $('[data-slider]').each(function () {
                             '<use xlink:href="/assets/svg/sprite.symbol.svg#arrow-down"></use>' +
                         '</svg>' +
                     '</a>'
+    })
+    .on('afterChange', function(event, slick) {
+        Tracking.track({
+            category: 'Slider',
+            action: 'Changed during user action',
+            label: 'Current Slide '.concat(slick.currentSlide)
+        });
     });
 });
+